requirements.txt
django==4.2.2
psycopg2==2.9.6
Pillow==9.5.0
django-environ==0.10.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.1
django-allauth==0.54.0
django-modeltranslation==0.18.10
